requirements.txt
jupyter
matplotlib
numpy
openai>=1.0.0
pandas
python-dotenv
scipy
